+++ Tests/Components/Renderers/TestLineRenderer.cs
@@ -1,5 +1,0 @@
-namespace Termule.Tests.Components;
-
-public class TestLineRenderer
-{
-}