+++ Tests/Core/TestGameElement.cs
@@ -1,6 +1,6 @@
 using Termule.Engine.Core;
 using Termule.Engine.Exceptions;
-using Termule.Tests.Utilities;
+using Termule.Tests.Core.Fakes;
 
 namespace Termule.Tests.Core;