+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Renamed `Frame` to `FrameBuffer` and moved it to the display namespace
 - Improved `FrameBuffer` performance by removing unnecessary delegate allocations
 - Made GameObjects only add components to the tick list on the next frame after they are added to avoid snapshot allocation
+- Rename `TransformRenderer` to `PositionalRenderer` and update the API
 
 ### Removed